+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update the process manual test admin API. [#114](https://github.com/rokwire/health-building-block/issues/114)
 
 ## [2.4.0] - 2021-01-11
 ### Security